Access the UART with 8-bit accesses instead of 32-bit, as this works on more platforms

+++ arch/arm/genmmserial-bs.c
@@ -15,7 +15,7 @@
 #include <stdint.h>
 #include "bootstrap.h"
 
-#define REGISTER(addr) (*(volatile uint32_t *)(addr))
+#define REGISTER(addr) (*(volatile uint8_t *)(addr))
 
 #define REG_THR REGISTER(&__dbg_serial_phys__)
 #define REG_LSR REGISTER(&__dbg_serial_phys__ + 0x14)
